+++ api/upscaler_tpu.py
@@ -1,4 +1,4 @@
-from pydantic import BaseModel, Field
+from fastapi import File, Form, UploadFile
 import base64
 from io import BytesIO
 from PIL import Image
@@ -6,6 +6,7 @@ from api.base_api import BaseAPIRouter, change_dir, init_helper
 from fastapi.responses import JSONResponse
 from fastapi.encoders import jsonable_encoder
 from repo.upscaler_tpu.pipeline import UpscaleModel
+from typing import Optional
 
 app_name = "upscaler_tpu"
 
@@ -22,19 +23,26 @@ class AppInitializationRouter(BaseAPIRouter):
 
 router = AppInitializationRouter(app_name=app_name)
 
-class UpscaleRequest(BaseModel):
-    image: str = Field(..., description="Base64 encoded image to upscale")
-    upscale_ratio: float = Field(..., description="Upscale ratio")
 
-@router.post("/upscale")
+### 图像超分；兼容openai api，image/variations
+@router.post("/v1/images/variations")
 @change_dir(router.dir)
-async def face_swap(request: UpscaleRequest):
-    src_image_bytes = BytesIO(base64.b64decode(request.image))
-    src_image = Image.open(src_image_bytes)
-    pil_res = router.models.extract_and_enhance_tiles(src_image, upscale_ratio=request.upscale_ratio)
+async def face_swap(
+    image: UploadFile = File(...),
+    upscale_ratio: Optional[float] = Form(1.0),
+):
+    ori_image_bytes = await image.read()
+    src_image = Image.open(BytesIO(ori_image_bytes))
+    pil_res = router.models.extract_and_enhance_tiles(src_image, upscale_ratio=upscale_ratio)
     # pil to base64
     buffer = BytesIO()
     pil_res.save(buffer, format='JPEG')
     ret_img_b64 = base64.b64encode(buffer.getvalue()).decode('utf-8')
-    content = {'ret_img': [ret_img_b64], 'message': 'success'}
+    content = {
+                "data": [
+                    {
+                        "b64_json": ret_img_b64
+                    }
+                        ]
+                }
     return JSONResponse(content=jsonable_encoder(content), media_type="application/json")
